alfred: Retrieve aliases from alfred and integrate batadv-vis.
Extending the visualisation data sources by the new batadv-vis,
which itself base ontop of alfred. Also extending the aliases import
by parsing informations from alfred. The last part is possible
through the preprocessing of data with alfred-json, so we don't
need to process the 'broken' format alfred itself supports.

This patch needs batadv-vis and tcatm/alfred-json in the PATH.